+++ examples/somersaultecu.py
@@ -462,6 +462,7 @@ somersault_dops = {
             unit_ref=None,
             sdgs=[],
             internal_constr=None,
+            physical_constr=None,
         ),
     "soberness_check":
         DataObjectProperty(
@@ -476,6 +477,7 @@ somersault_dops = {
             unit_ref=None,
             sdgs=[],
             internal_constr=None,
+            physical_constr=None,
         ),
     "dizzyness_level":
         DataObjectProperty(
@@ -490,6 +492,7 @@ somersault_dops = {
             unit_ref=None,
             sdgs=[],
             internal_constr=None,
+            physical_constr=None,
         ),
     "happiness_level":
         DataObjectProperty(
@@ -504,6 +507,7 @@ somersault_dops = {
             unit_ref=None,
             sdgs=[],
             internal_constr=None,
+            physical_constr=None,
         ),
     "duration":
         DataObjectProperty(
@@ -518,6 +522,7 @@ somersault_dops = {
             unit_ref=OdxLinkRef.from_id(somersault_units["second"].odx_id),
             sdgs=[],
             internal_constr=None,
+            physical_constr=None,
         ),
     "temperature":
         DataObjectProperty(
@@ -532,6 +537,7 @@ somersault_dops = {
             unit_ref=OdxLinkRef.from_id(somersault_units["celsius"].odx_id),
             sdgs=[],
             internal_constr=None,
+            physical_constr=None,
         ),
     "error_code":
         DataObjectProperty(
@@ -546,6 +552,7 @@ somersault_dops = {
             unit_ref=None,
             sdgs=[],
             internal_constr=None,
+            physical_constr=None,
         ),
     "boolean":
         DataObjectProperty(
@@ -561,6 +568,7 @@ somersault_dops = {
             unit_ref=None,
             sdgs=[],
             internal_constr=None,
+            physical_constr=None,
         ),
     "uint8":
         DataObjectProperty(
@@ -575,6 +583,7 @@ somersault_dops = {
             unit_ref=None,
             sdgs=[],
             internal_constr=None,
+            physical_constr=None,
         ),
     "float":
         DataObjectProperty(
@@ -589,6 +598,7 @@ somersault_dops = {
             unit_ref=None,
             sdgs=[],
             internal_constr=None,
+            physical_constr=None,
         ),
 }
