+++ tests/test_singleecujob.py
@@ -110,6 +110,7 @@ class TestSingleEcuJob(unittest.TestCase):
                 unit_ref=None,
                 sdgs=[],
                 internal_constr=None,
+                physical_constr=None,
             ),
             outputDOP=DataObjectProperty(
                 odx_id=OdxLinkId("ID.outputDOP", doc_frags),
@@ -139,6 +140,7 @@ class TestSingleEcuJob(unittest.TestCase):
                 unit_ref=None,
                 sdgs=[],
                 internal_constr=None,
+                physical_constr=None,
             ),
             negOutputDOP=DataObjectProperty(
                 odx_id=OdxLinkId("ID.negOutputDOP", doc_frags),
@@ -168,6 +170,7 @@ class TestSingleEcuJob(unittest.TestCase):
                 unit_ref=None,
                 sdgs=[],
                 internal_constr=None,
+                physical_constr=None,
             ),
         )
